Pull data through API and routes
We now have our data mapped through the view-section component
into the section component, we do this by OnInit of view-section
performing our api call and storing the current tutorial in the
Tutorial service. When we go to a section, we take that section name
and pass it as route params to our auxillary route. Upon the init
for the section we pull data back down from the Tutorial service,
our service has a method which maps the observable for our
current section.